protectong routes with authorization

+++ backend/routes/posts.js
@@ -106,18 +106,24 @@ router.put(
       content: req.body.content,
       imagePath: imagePath,
     });
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      console.log(result);
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+      if(result.nModified > 0){
+        res.status(200).json({ message: "Update successful!" });
+      }else {
+        res.status(401).json({ message: "Not Authorized!" });
+      }
     });
   }
 );
 
 //Delete Post
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if(result.n > 0){
+      res.status(200).json({ message: "Deleted Successfully!" });
+    }else {
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   });
 });
